% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkCapitalization}
\alias{checkCapitalization}
\title{Helper function for checking Capitalization}
\usage{
checkCapitalization(ds, var_list)
}
\arguments{
\item{ds}{dataset to check}

\item{var_list}{vector of variables to check}
}
\value{
nothing
}
\description{
Helper function for checking Capitalization
}

