% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hormHelpers.R
\name{checkPlotArea}
\alias{checkPlotArea}
\title{Helper function for checking plot options for hormArea}
\usage{
checkPlotArea(m, l_b)
}
\arguments{
\item{l_b}{lower_bound}

\item{d}{date_format}
}
\value{
none
}
\description{
Helper function for checking plot options for hormArea
}

