% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkPlotOpts}
\alias{checkPlotOpts}
\title{Helper function for checking plot options}
\usage{
checkPlotOpts(p_p_p, w, h, s, x = NA, y = NA, d)
}
\arguments{
\item{w}{plot_width}

\item{h}{plot_height}

\item{s}{save_plot}

\item{x}{xscale}

\item{y}{yscale}

\item{d}{date_format}

\item{p_p}{plot_per_page}

\item{l_b}{lower_bound}
}
\value{
none
}
\description{
Helper function for checking plot options
}

