% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkPlotOverlap}
\alias{checkPlotOverlap}
\title{Helper function for checking plot options for hormPlotOverlap}
\usage{
checkPlotOverlap(h_v, b_v)
}
\arguments{
\item{h_v}{hormone_var}

\item{b_v}{by_var}
}
\value{
none
}
\description{
Helper function for checking plot options for hormPlotOverlap
}

