% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hormHelpers.R
\name{checkPlotRatio}
\alias{checkPlotRatio}
\title{Helper function for checking plot options for hormPlotOverlap}
\usage{
checkPlotRatio(x, h_v, h_n, h_d, b_v)
}
\arguments{
\item{h_v}{hormone_var}

\item{b_v}{by_var}
}
\value{
none
}
\description{
Helper function for checking plot options for hormPlotOverlap
}

