% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hormHelpers.R
\name{cleanByvar}
\alias{cleanByvar}
\title{Splits and trims by_var}
\usage{
cleanByvar(x)
}
\arguments{
\item{x}{by_var to be cleans (e.g. 'species, hormone' )}
}
\value{
x split and trimmed of whitespace
}
\description{
Splits and trims by_var
}
\examples{
ds <- hormRead()
}

