% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hormHelpers.R
\name{getAddLine}
\alias{getAddLine}
\title{Helper function for getting Baseline for plotting}
\usage{
getAddLine(d_s, crit, conc, a_l)
}
\arguments{
\item{d_s}{ds_sub}

\item{crit}{x$criteria}

\item{conc}{conc_var}

\item{a_l}{add_line (which line type to return)}
}
\value{
baseline
}
\description{
Helper function for getting Baseline for plotting
}

