% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hormHelpers.R
\name{getAUC}
\alias{getAUC}
\title{Get the area under a curve using trapezoid method}
\usage{
getAUC(t, c, m)
}
\arguments{
\item{t}{time vector}

\item{c}{response vector (e.g. hormone concentration)}

\item{m}{method for calculating AUC (trapezoid, spline)}
}
\value{
calculates the area under the curve for the curve
}
\description{
Get the area under a curve using trapezoid method
}
\examples{
time <- 1:10
conc <- c( rep(0,4),2,2,rep(0,4))
plot(conc~time,type='l')
getAUC(time,conc,'trapezoid')
getAUC(time,conc,'spline')
}

