% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getCutoff}
\alias{getCutoff}
\title{Helper function to calculate baseline cutoff}
\usage{
getCutoff(x, criteria)
}
\arguments{
\item{x}{hormone concentration}

\item{criteria}{the number of standard deviations above baseline}
}
\value{
cutoff value
}
\description{
Helper function to calculate baseline cutoff
}
\examples{
conc <- rnorm(100)
getCutoff(conc, 2)
}

