% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hormHelpers.R
\name{getEventInfo}
\alias{getEventInfo}
\title{Helper function for getting event info}
\usage{
getEventInfo(d_s, e, t)
}
\arguments{
\item{d_s}{ds_sub}

\item{e}{x$event_var}

\item{t}{time_var}
}
\value{
events
}
\description{
Helper function for getting event info
}

