% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hormHelpers.R
\name{getPeakInfo}
\alias{getPeakInfo}
\title{Assigns peak number and get area under curve for each peak}
\usage{
getPeakInfo(k, date = time_var, conc = conc_var, m_pk)
}
\arguments{
\item{k}{dataframe (broken first by by_var)}

\item{date}{time variable}

\item{conc}{response variable (e.g. concentration)}

\item{m_pk}{method used to calculate peak}
}
\value{
return dataset with peaks identified and AUC calculated
}
\description{
Assigns peak number and get area under curve for each peak
}
\examples{
head(hormone)
}

