% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getPlotlim}
\alias{getPlotlim}
\title{Helper function for get axes limits}
\usage{
getPlotlim(d_s, d_f, var, scale, max_expand = 1, base = NA)
}
\arguments{
\item{d_s}{ds_sub dataframe}

\item{d_f}{data dataframe}

\item{var}{response variable}

\item{scale}{xscale or yscale}

\item{base}{baseline measure which is needed if higher than max(data[,conc_var])}
}
\value{
limits a vector of min and max
}
\description{
Helper function for get axes limits
}

