% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getPlotTitle}
\alias{getPlotTitle}
\title{Combined the columns listed in the by_var_v}
\usage{
getPlotTitle(data, by_var = by_var_v)
}
\arguments{
\item{data}{dataframe containing the variables in by_var_v}

\item{by_var}{the by_var_v (vector of names)}
}
\value{
the combined columns as a single vector, separated by a semicolon
}
\description{
Combined the columns listed in the by_var_v
}
\examples{
(plot_title <- getPlotTitle(hormone,by_var=c('sp','sex')))
}

