% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSumStat}
\alias{getSumStat}
\title{Helper function to get aggregate statistics}
\usage{
getSumStat(data = ds, name = "mean", func = function(x) mean(x, na.rm =
  T), add_ds, c_var = conc_var, by_var = by_var_v)
}
\arguments{
\item{data}{dataframe}

\item{name}{new name for variable being produced}

\item{func}{function to be used in aggregate()}

\item{add_ds}{optional dataframe - if included, the new variable will be added to the supplied dataframe}

\item{c_var}{name of the conc_var (e.g. concentration)}

\item{by_var}{vector of by_var (usually use by_var_v )}
}
\value{
return either new variable by itself or in the dataframe listed in add_ds
}
\description{
Helper function to get aggregate statistics
}
\examples{
ds <- getSumStat(hormone, name='mean_var', func= function(x) mean(x,na.rm=T), by_var=c('sp','sex'), c_var='conc' )
ds
}

