% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hormACTH}
\alias{hormACTH}
\title{Run iterative process to calculate baseline.}
\usage{
hormACTH(data)
}
\arguments{
\item{data}{dataset.}

\item{by_var}{grouping variable.}

\item{time_var}{grouping variable.}

\item{conc_var}{grouping variable.}
}
\value{
hormLong object.
}
\description{
Run iterative process to calculate baseline.
}
\examples{
result <- hormACTH(data=ACTH)
}

