% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hormAuto}
\alias{hormAuto}
\title{Automated function that runs baseline calculation, make plots, outputs dataset as csv}
\usage{
hormAuto(file_method = "write", by_var, time_var, conc_var,
  filename = "test")
}
\arguments{
\item{file_method}{if 'choose' then user is prompted. if 'write' then user writes the name}
}
\value{
nothing in R.  Output hormPlot.pdf and horm_data_out.csv
}
\description{
Automated function that runs baseline calculation, make plots, outputs dataset as csv
}
\examples{
ds <- hormAuto(by_var='sp, sex, id', time_var='date', conc_var='conc')
}

