% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hormCutoff}
\alias{hormCutoff}
\title{Helper function to calculate baseline cutoff}
\usage{
hormCutoff(x, criteria)
}
\arguments{
\item{x}{hormone concentration}

\item{criteria}{the number of standard deviations above baseline}
}
\value{
nothing
}
\description{
Helper function to calculate baseline cutoff
}
\examples{
result <- hormBaseline(data=hormone, by_var='sp, sex, id', time_var='date', conc_var='conc' )
hormPlot(result)
hormWrite(result, file_out='data_out.csv', file_type='csv' )
}

