% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{hormLynx}
\alias{hormLynx}
\title{Canadian Lynx data}
\format{A data frame with five variables: \code{AnimalID}, \code{SampleNo},
  \code{Date},\code{Time}, \code{Hormone}, \code{Conc}, \code{Events}, and \code{datetime}}
\usage{
hormLynx
}
\description{
Longitudinal hormone data for cort for multiple lynx individuals
}
\details{
Reference: Fanson KV,  Wielebnowski NC, Shenk TM, Lucas JR. Comparative patterns of adrenal activity in
captive and wild Canada lynx (Lynx canadensis). Journal of Comparative Physiology B 182:157-165
}
\keyword{datasets}

