% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{hormone2}
\alias{hormone2}
\title{Hormone2 data.}
\format{A data frame with five variables: \code{sp}, \code{sex},
  \code{id}, \code{date} and \code{conc}}
\usage{
hormone2
}
\description{
Longitudinal hormone data for three species, both sexes, and multiple
individuals
}
\keyword{datasets}

