% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hormPlotOverlap}
\alias{hormPlotOverlap}
\title{Plot overlapping longitudinal hormone data}
\usage{
hormPlotOverlap(x, hormone_var = "horm_type", date_format = "\%d-\%b",
  colors = c("red", "blue"), plot_per_page = 4, save_plot = TRUE,
  plot_height = 2, plot_width = 6, yscale = "free", xscale = "free",
  ...)
}
\arguments{
\item{x}{hormLong object (produced from hormBaseline) [required]}

\item{hormone_var}{name for the hormone variable.  It must be listed in by_var during hormBaseline [required]}

\item{date_format}{the format of the date variable on x-axis. See help for examples of other formats [default = '\%d-\%b']}

\item{colors}{vector of colors for the fills. It needs to have as many colors as hormone types  [default=c('red','blue')]}

\item{plot_per_page}{the number of plot panels per page, by row. [default = 4]}

\item{save_plot}{indicates whether to save plot as a file [default = TRUE]}

\item{plot_height}{the height of individual plot panels (in inches).  Pdf page height is determined by both plot_per_page and plot_height. [default = 2]}

\item{plot_width}{the width of the pdf page. [default = 6]}

\item{yscale}{determines if y-axis should be free ('free') to change for each panel or remain the same ('fixed') for all panels [default = 'free']}

\item{xscale}{determines if x-axis should be free ('free') to change for each panel or remain the same ('fixed') for all panels  [default = 'free']}

\item{...}{generic plotting options [optional]}
}
\value{
nothing  Produces a pdf file saved at current working directory
}
\description{
Plot overlapping longitudinal hormone data
}
\examples{
result <- hormBaseline(data=hormLynx, criteria=2, by_var='AnimalID, Hormone', time_var='Date', conc_var='Conc' )
hormPlotOverlap( result, hormone_var='Hormone', colors=c('red','blue','yellow') )
}

