% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hormPlotRatio}
\alias{hormPlotRatio}
\title{Plot longitudinal hormone ratios}
\usage{
hormPlotRatio(x, horm_var, horm_num, horm_denom, ...)
}
\arguments{
\item{x}{hormLong object (produced from hormBaseline) [required]}

\item{horm_var}{name of the hormone variable.  This must have been included
as part of by_var in hormBaseline  [required]}

\item{horm_num}{name of the hormone that is the numerator in the ratio. [required]}

\item{horm_denom}{name of the hormone that is the denominator in the ratio. [required]}

\item{...}{arguments for hormPlot [optional]}
}
\value{
nothing  Produces a pdf file saved at current working directory
}
\description{
Plot longitudinal hormone ratios
}
\examples{
ds <- hormDate(hormone2,date_var = 'date', date_order = 'ymd')
res <- hormBaseline(data=ds,criteria=1,by_var='sp,horm_type,id',conc_var = 'conc',time_var='date',
                     event_var='event')
hormPlotRatio( x=res, horm_var='horm_type', horm_num='E', horm_denom='prog' )
}

