% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hormHelpers.R
\name{hormRead}
\alias{hormRead}
\title{Read hormone data from a csv file}
\usage{
hormRead()
}
\arguments{
\item{none}{no arguments}
}
\value{
imported dataset
}
\description{
Read hormone data from a csv file
}
\examples{
ds <- hormRead()
}

