% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hormSmooth}
\alias{hormSmooth}
\title{Create longitudinal plot with smoothers}
\usage{
hormSmooth(data, by_var, group_var, time_var, conc_var, plot_per_page = 4,
  plot_height = 2, plot_width = 6, yscale = "free", xscale = "free",
  smoothness = c(0.7, 0.7), shape = c(19, 20), colour = c("red", "blue"),
  line_type = c(1, 2), line_width = c(1, 1), ...)
}
\arguments{
\item{data}{dataset.}

\item{by_var}{by variable.}

\item{group_var}{grouping variable.}

\item{time_var}{time variable.}

\item{conc_var}{response variable.}
}
\value{
hormLong object.
}
\description{
Create longitudinal plot with smoothers
}
\examples{
hormSmooth(data=hormone2, by_var='sp, id', group_var='horm_type', time_var='date', conc_var='conc')
}

