% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotAxes}
\alias{plotAxes}
\title{Helper function for get axes limits}
\usage{
plotAxes(d_s, t, x_s, d_f)
}
\arguments{
\item{d_s}{ds_sub dataframe}

\item{d_f}{data dataframe}

\item{var}{response variable}

\item{scale}{xscale or yscale}

\item{base}{baseline measure which is needed if higher than max(data[,conc_var])}
}
\value{
limits a vector of min and max
}
\description{
Helper function for get axes limits
}

