% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotEventInfo}
\alias{plotEventInfo}
\title{Helper function that plots Event info}
\usage{
plotEventInfo(d_e = events, t = x$time_var, e = x$event_var)
}
\arguments{
\item{d_e}{events dataframe}

\item{t}{time_var name}

\item{e}{event_var name}
}
\value{
nothing plots event text on graphs using par('user)[4] coordinate
}
\description{
Helper function that plots Event info
}

