% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ridFactor}
\alias{ridFactor}
\title{Helper function that converts all factors to string in a dataset}
\usage{
ridFactor(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
dataframe with all factors converted to strings
}
\description{
Helper function that converts all factors to string in a dataset
}
\examples{
sapply(iris,class)
ds <- ridFactor(iris)
sapply(ds,class)
}

