% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write.rtf}
\alias{write.rtf}
\title{Write results to rtf file}
\usage{
write.rtf(x, filename)
}
\arguments{
\item{x}{dataframe to be writen to a file}

\item{file}{filename}
}
\value{
x split and trimmed of whitespace
}
\description{
Write results to rtf file
}
\examples{
write.rtf(cars,'cars.rtf')
}

