% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hormHelpers.R
\name{checkClass}
\alias{checkClass}
\title{Helper function for checking class}
\usage{
checkClass(var, v_class)
}
\arguments{
\item{var}{ds_sub dataframe}

\item{class}{baseline measure which is needed if higher than max(data[,conc_var])}
}
\value{
none
}
\description{
Helper function for checking class
}

