% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hormHelpers.R
\name{checkPlotBreaks}
\alias{checkPlotBreaks}
\title{Helper function for checking plot options for hormPlotBreaks}
\usage{
checkPlotBreaks(b_c, b_b)
}
\arguments{
\item{b_c}{break_cutoff}

\item{b_b}{break_buffer}
}
\value{
none
}
\description{
Helper function for checking plot options for hormPlotBreaks
}

