% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkPlotMissing}
\alias{checkPlotMissing}
\title{Helper function for checking for missing values in data before plotting}
\usage{
checkPlotMissing(ds, var_list, conc)
}
\arguments{
\item{var_list}{vector of variables to check}

\item{conc}{conc_var (string)}

\item{data}{dataset to check}
}
\value{
dataframe
}
\description{
Helper function for checking for missing values in data before plotting
}

