% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hormHelpers.R
\name{checkPlotOverlap}
\alias{checkPlotOverlap}
\title{Helper function for checking plot options for hormPlotOverlap}
\usage{
checkPlotOverlap(h_v, b_v)
}
\arguments{
\item{h_v}{hormone_var}

\item{b_v}{by_var}
}
\value{
none
}
\description{
Helper function for checking plot options for hormPlotOverlap
}

