% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getBaseline}
\alias{getBaseline}
\title{Helper function for getting Baseline for plotting}
\usage{
getBaseline(d_s, crit, conc)
}
\arguments{
\item{d_s}{ds_sub}

\item{crit}{x$criteria}

\item{conc}{conc_var}
}
\value{
baseline
}
\description{
Helper function for getting Baseline for plotting
}

