% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hormArea.R
\name{hormArea}
\alias{hormArea}
\title{Plot the area under the curve (AUC)}
\usage{
hormArea(x, lower_bound = "origin", method = "trapezoid",
  date_format = "\%d-\%b", y_label = NA, xscale = "free",
  yscale = "free", plot_per_page = 4, plot_height = 2, plot_width = 6,
  save_plot = T)
}
\arguments{
\item{x}{hormLong object (produced from hormBaseline) [required]}

\item{lower_bound}{the lower limit to calculate the area under the curve. This can be
'origin', 'baseline', or 'peak'.  Origin is all values above 0.  Baseline is all values above
baseline mean.  Peak is all values above peak cutoff. [default = 'origin']}

\item{method}{the AUC method to use. Only Trapezoid and spline method has been implemented. Units
for AUC is conc units * day. [default = 'trapezoid']}

\item{date_format}{the format of the date variable on x-axis. Default is 01-Jan format. See Appendix 1 in help manual
for examples of other formats [default = '\%d-\%b']}

\item{y_label}{sets the y-axis label.  The default setting is the column name of the response variable. [default = NA]}

\item{xscale}{determines if x-axis is free ('free') to change for each panel or
remain the same ('fixed') for all panels  [default = 'free']}

\item{yscale}{determines if y-axis is free ('free') to change for each panel or
remain the same ('fixed') for all panels  [default = 'free']}

\item{plot_per_page}{the number of plot panels per page, by row. [default = 4]}

\item{plot_height}{the height of individual plot panels (in inches).  Pdf page height is determined
by both plot_per_page and plot_height. [default = 2]}

\item{plot_width}{the width of the pdf page. [default = 6]}

\item{save_plot}{indicates whether [default = TRUE]}
}
\value{
nothing is returned.  Pdf file is produced and a summary table
}
\description{
Plot the area under the curve (AUC)
}
\details{
Citation for the AUC used in this example:\cr\cr
  Cockrem JF and Silverin B. 2002. Variation within and between Birds in Corticosterone
  Responses of Great Tits (Parus major).  General and Comparative Endocrinology 125:197.
}
\examples{
result <- hormBaseline(data=hormLynx, criteria=2, by_var='AnimalID, Hormone',
            time_var='datetime', conc_var='Conc', event_var='Events' )

# AUC only for values above cutoff threshold
#  defined in hormBaseline
hormArea(result,lower_bound='peak')

# AUC only for values above baseline mean
hormArea(result,lower_bound='baseline')

# AUC only for values above 0
hormArea(result,lower_bound='origin')
}

