% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hormAuto.R
\name{hormAuto}
\alias{hormAuto}
\title{Automated function that runs baseline calculation, make plots, outputs dataset as csv}
\usage{
hormAuto(by_var, time_var, conc_var, filename = "test")
}
\arguments{
\item{none}{}
}
\value{
nothing in R.  Output hormPlot.pdf and horm_data_out.csv
}
\description{
Automated function that runs baseline calculation, make plots, outputs dataset as csv
}
\examples{
ds <- hormAuto(by_var='sp, sex, id', time_var='date', conc_var='conc')
}

