% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hormBaseline}
\alias{hormBaseline}
\title{Run iterative process to calculate baseline.}
\usage{
hormBaseline(data, by_var, criteria = 2, conc_var, time_var)
}
\arguments{
\item{data}{dataset.}

\item{grp}{grouping variable.}
}
\value{
sumarised dataset.
}
\description{
Run iterative process to calculate baseline.
}
\examples{
result <- hormBaseline(data=hormone, by_var='sp, sex, id', time_var='date', conc_var='conc' )
}

