% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hormBaseline}
\alias{hormBaseline}
\title{Run iterative process to calculate baseline.}
\usage{
hormBaseline(data, by_var, conc_var, time_var, criteria = 2)
}
\arguments{
\item{data}{dataset.}

\item{by_var}{grouping variable.}

\item{conc_var}{grouping variable.}

\item{time_var}{grouping variable.}

\item{criteria}{grouping variable.}
}
\value{
hormLong object.
}
\description{
Run iterative process to calculate baseline.
}
\examples{
result <- hormBaseline(data=hormone, by_var='sp, sex, id', time_var='date', conc_var='conc' )
}

