% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hormBoxplot}
\alias{hormBoxplot}
\title{Boxplot of individual concentrations}
\usage{
hormBoxplot(data, conc_var, id_var, plot_height = 4, plot_width = 6,
  save_plot = T)
}
\arguments{
\item{data}{dataset}

\item{id_var}{name of the individual variable}
}
\value{
NA
}
\description{
Boxplot of individual concentrations
}
\examples{
hormBoxplot(data=hormone, conc_var='conc',id_var='id')
}

