% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hormCutoff}
\alias{hormCutoff}
\title{Helper function to calculate baseline cutoff}
\usage{
hormCutoff(x, criteria)
}
\arguments{
\item{x}{hormone concentration}

\item{criteria}{the number of standard deviations above baseline}
}
\value{
nothing
}
\description{
Helper function to calculate baseline cutoff
}
\examples{
conc <- rnorm(100)
hormCutoff(conc, 2)
}

