% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hormDate}
\alias{hormDate}
\title{Helper function to work with dates}
\usage{
hormDate(data, date_var, time_var)
}
\arguments{
\item{data}{data set with the variables to convert}

\item{date_var}{variable name for date column in format '2014-jan-01'}

\item{time_var}{variable name for the time column in format '10:10 PM'}
}
\value{
new datetime variable
}
\description{
Helper function to work with dates
}
\examples{
date <- c('2014-jan-01','2014-jan-02')
time <- c('10:10 Am', '10:30 pm')
hormDate(date,time)
}

