% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hormDate}
\alias{hormDate}
\title{Helper function to work with dates.  User needs to specify at least date_var and date_order.
If time_var is included, then it will output a date_time variable}
\usage{
hormDate(data, date_var, time_var, name = "datetime", date_order = "ymd")
}
\arguments{
\item{data}{data set with the variables to convert [required]}

\item{date_var}{variable name for date column.  Form should be a standard date form
(e.g. '12AUG2014','01/01/2014') [required]}

\item{time_var}{variable name for the time column in 24-hour format (e.g. '20:10:01', '01:04:01') [optional]}

\item{name}{name of the new date variable created by the function [default = 'datetime']}

\item{date_order}{specifies the day, month, year order in abbreviated form (e.g. 'ymd','mdy','dmy').
[defualt='ymd']}
}
\value{
dataset with the new date/time variable
}
\description{
Helper function to work with dates.  User needs to specify at least date_var and date_order.
If time_var is included, then it will output a date_time variable
}
\examples{
date <- c('2014-01-01','2014-01-01')
time <- c('10:10:01', '20:30:23')
ds <- data.frame(date=date,time=time)
hormDate(ds,date_var='date',time_var='time',name='datetime', date_order='ymd' )
}

