% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{hormElephant}
\alias{hormElephant}
\title{Asian Elephant data}
\format{A data frame with five variables: \code{Ele},
  \code{Date_collected}, \code{Hormone}, \code{Conc_ng_ml} and \code{Event}}
\usage{
hormElephant
}
\description{
Longitudinal hormone data for multiple female elephants in which LH and Progesterone were
measured over years.
}
\details{
Reference: Fanson KV, Keeley T, Fanson BG. Cyclic changes in cortisol across the estrous
cycle in parous and nulliparous Asian elephants. Endocrine Connections 3:57-66
}
\keyword{datasets}

