% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hormPlot.R
\name{hormPlot}
\alias{hormPlot}
\title{Plot longitudinal hormone data with baseline information}
\usage{
hormPlot(x, add_line = "baseline_cutoff", date_format = "\%d-\%b",
  break_cutoff = Inf, color = "black", symbol = 19, y_label = NA,
  xscale = "free", yscale = "free", plot_per_page = 4, plot_height = 2,
  plot_width = 6, filename = "hormPlot", save_plot = TRUE)
}
\arguments{
\item{x}{hormLong object (produced from hormBaseline) [required]}

\item{add_line}{adds a horizontal line to all plots representing baseline mean,
baseline cutoff (mean + critera*SD), or mean (all values). Options are 'baseline_mean', 'baseline_cutoff',
'mean' or 'none' [default = 'baseline_cutoff']}

\item{date_format}{the format of the date variable on x-axis.  Default is 01-Jan format. See Appendix 1 in help manual
for examples of other formats [default = '\%d-\%b']}

\item{break_cutoff}{the maximum number of days between consecutive points that are
still joined by a line. Default is to connect all points with a line. [default = Inf]}

\item{color}{color of the line and points [default='black']}

\item{symbol}{number to indicate point symbol. e.g. 1=open circle, 2=open triangle, 15=closed square, 19=closed circle  [default=19]}

\item{y_label}{sets the y-axis label.  The default setting is the column name of the response variable. [default = NA]}

\item{xscale}{determines if x-axis is free ('free') to change for each panel or remain the same ('fixed') for all panels  [default = 'free']}

\item{yscale}{determines if y-axis is free ('free') to change for each panel or remain the same ('fixed') for all panels [default = 'free']}

\item{plot_per_page}{the number of plot panels per page, by row. [default = 4]}

\item{plot_height}{the height of individual plot panels (in inches).
Pdf page height is determined by both plot_per_page and plot_height. [default = 2]}

\item{plot_width}{the width of the pdf page. [default = 6]}

\item{filename}{the filename of the pdf file [default = 'hormPlot']}

\item{save_plot}{indicates whether to save plot as a file [default = TRUE]}

\item{file_type}{determines the file type for outputted plots [default = TRUE]}
}
\value{
nothing  Produces a pdf file saved at current working directory
}
\description{
Plot longitudinal hormone data with baseline information
}
\examples{
result <- hormBaseline(data=hormLynx, criteria=2, by_var='AnimalID, Hormone', time_var='datetime',
                       conc_var='Conc', event_var='Events' )
hormPlot( result ) # default settings...adds a line for baseline_cutoff

# if you want to add lines for none, baseline_mean or overall mean
hormPlot( result, add_line='none' ) # no line
hormPlot( result, add_line='mean' ) # adds a line for individual mean
hormPlot( result, add_line='baseline_mean' ) # adds a line for mean of just baseline values

# Sometimes you may want to fix y-axis to be same across all plots so heights are comparable
hormPlot( result, yscale='fixed', color='red', symbol=1 )

# You may want to just include both date and time on the x-axis instead of date
hormPlot( result, date_format='\%d/\%m/\%y \%H:\%m' )   # does not makes sense here but the code shows how


# You may want to not join points that are temporally separated by large gaps
result <- hormBaseline(data=hormElephant, criteria=2, by_var='Ele, Hormone', time_var='Date',
             conc_var='Conc_ng_ml' , event_var='Event')
hormPlot( result, y_label='Conc (ng/ml)' )  # everything connected
hormPlot( result, yscale='fixed', break_cutoff=30 ) # do not join any points more than 20 days
}

