% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hormPlot}
\alias{hormPlot}
\title{Plot longitudinal hormone data with baseline information}
\usage{
hormPlot(x, plot_per_page = 4, save_plot = TRUE, plot_height = 2,
  plot_width = 6, ...)
}
\arguments{
\item{x}{hormLong object (produced from hormBaseline)}

\item{plot_per_page}{the number of plot panels per page, by row.}

\item{save_plot}{indicates whether}

\item{plot_height}{the height of individual plot panels (in inches).  Pdf page height is determined
by both plot_per_page and plot_height.}

\item{plot_width}{the width of the pdf page.}
}
\value{
NA
}
\description{
Plot longitudinal hormone data with baseline information
}
\examples{
result <- hormBaseline(data=hormone, by_var='sp, sex, id', time_var='date', conc_var='conc' )
hormPlot(result)
}

