% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hormPlot}
\alias{hormPlot}
\title{Plot longitudinal hormone data with baseline information}
\usage{
hormPlot(x, plot_per_page = 4, save_plot = TRUE, plot_height = 2,
  plot_width = 6, yscale = "free", xscale = "free", ...)
}
\arguments{
\item{x}{hormLong object (produced from hormBaseline) [required]}

\item{plot_per_page}{the number of plot panels per page, by row. [default = 4]}

\item{save_plot}{indicates whether to save plot as a file [default = TRUE]}

\item{plot_height}{the height of individual plot panels (in inches).  Pdf page height is determined
by both plot_per_page and plot_height. [default = 2]}

\item{plot_width}{the width of the pdf page. [default = 6]}

\item{yscale}{determines if y-axis should be free ('free') to change for each panel or
remain the same ('fixed') for all panels [default = 'free']}

\item{xscale}{determines if x-axis should be free ('free') to change for each panel or
remain the same ('fixed') for all panels  [default = 'free']}
}
\value{
nothing  Produces a pdf file saved at current working directory
}
\description{
Plot longitudinal hormone data with baseline information
}
\examples{
result <- hormBaseline(data=hormone, criteria=1, by_var='sp, sex, id', time_var='date', conc_var='conc' )
hormPlot(result, yscale='fixed',xscale='fixed' )
}

