% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hormPlotBreaks}
\alias{hormPlotBreaks}
\title{Longitudinal graph with breaks}
\usage{
hormPlotBreaks(x, break_cutoff = 40, break_buffer = 60, plot_per_page = 4,
  save_plot = TRUE, plot_height = 2, plot_width = 6)
}
\arguments{
\item{x}{hormLong object (produced from hormBaseline)}

\item{plot_per_page}{the number of plot panels per page, by row.}

\item{save_plot}{indicates whether}

\item{plot_height}{the height of individual plot panels (in inches).  Pdf page height is determined
by both plot_per_page and plot_height.}

\item{plot_width}{the width of the pdf page.}

\item{yscale}{determines if y-axis should be free ('free') to change for each panel or
remain the same ('fixed') for all panels}

\item{xscale}{determines if x-axis should be free ('free') to change for each panel or
remain the same ('fixed') for all panels}
}
\value{
NA
}
\description{
Longitudinal graph with breaks
}

