% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hormPlotBreaks.R
\name{hormPlotBreaks}
\alias{hormPlotBreaks}
\title{Longitudinal graph with breaks}
\usage{
hormPlotBreaks(x, break_cutoff = 40, break_buffer = 60,
  date_format = "\%d-\%b", color = "black", symbol = 19,
  plot_per_page = 4, plot_height = 2, plot_width = 6, save_plot = TRUE)
}
\arguments{
\item{x}{hormLong object (produced from hormBaseline) [required]}

\item{break_cutoff}{the maximum number of days between consecutive points.
Above this cutoff value, a break is created. [default = 40]}

\item{break_buffer}{size of the gap (in number of days) between data groups.  Larger values will
create larger spaces between data groups. [default = 60]}

\item{date_format}{the format of the date variable on x-axis. Default is 01-Jan format. See Appendix 1 in help manual
for examples of other formats [default = '\%d-\%b']}

\item{color}{colour of the line and points [default='black']}

\item{symbol}{number to indicate point symbol. e.g. 1=open circle, 2=open triangle, 15=closed square, 19=closed circle  [default=19]}

\item{plot_per_page}{the number of plot panels per page, by row. [default = 4]}

\item{plot_height}{the height of individual plot panels (in inches).  Pdf page height is determined
by both plot_per_page and plot_height. [default = 2]}

\item{plot_width}{the width of the pdf page.[default = 6]}

\item{save_plot}{indicates whether to save plot as a file [default = TRUE]}
}
\value{
nothing  Produces a pdf file saved at current working directory
}
\description{
Longitudinal graph with breaks
}
\examples{
ds <- hormElephant
result <- hormBaseline(data=ds, criteria=2, by_var='Ele, Hormone', time_var='Date',
             conc_var='Conc_ng_ml' , event_var='Event')
hormPlotBreaks( result )
hormPlotBreaks( result, color='red', symbol=1 )

# compare to regular hormPlot, especially Ele1; Cortisol
hormPlot( result )
}

