% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hormPlotRatio}
\alias{hormPlotRatio}
\title{Plot longitudinal hormone ratios}
\usage{
hormPlotRatio(x, hormone_var, hormone_num, hormone_denom, ...)
}
\arguments{
\item{x}{hormLong object (produced from hormBaseline) [required]}

\item{hormone_var}{name of the hormone variable.  This must have been included
as part of by_var in hormBaseline  [required]}

\item{hormone_num}{name of the hormone that is the numerator in the ratio. [required]}

\item{hormone_denom}{name of the hormone that is the denominator in the ratio. [required]}

\item{...}{arguments for hormPlot (e.g. break_cutoff, save_plot) [optional]}
}
\value{
nothing  Produces a pdf file saved at current working directory
}
\description{
Plot longitudinal hormone ratios
}
\examples{
result <- hormBaseline(data=hormElephant, criteria=2, by_var='Ele, Hormone', time_var='Date',
                       conc_var='Conc_ng_ml', event_var='Event' )
hormPlotRatio( x=result, hormone_var='Hormone', hormone_num='Cortisol', hormone_denom='Progesterone' )

# If you want to prevent points with large temporal gaps  joining, set the break_buffer (see hormPlot() )
hormPlotRatio( x=result, hormone_var='Hormone',
                         hormone_num='Cortisol',
                         hormone_denom='Progesterone',
                         break_cutoff=40, save_plot=F )
}

