% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hormSmooth}
\alias{hormSmooth}
\title{Create longitudinal plot with smoothers}
\usage{
hormSmooth(data)
}
\arguments{
\item{data}{dataset.}

\item{by_var}{by variable.}

\item{time_var}{time variable.}

\item{conc_var}{response variable.}

\item{group_var}{grouping variable.}
}
\value{
hormLong object.
}
\description{
Create longitudinal plot with smoothers
}
\examples{
result <- hormSmooth(data=ACTH)
}

