% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hormSmooth}
\alias{hormSmooth}
\title{Create longitudinal plot with smoothers (under development)}
\usage{
hormSmooth(data, by_var, group_var, time_var, conc_var, plot_per_page = 4,
  plot_height = 2, plot_width = 6, yscale = "free", xscale = "free",
  smoothness = 0.7, shape = 1, colour = "black", line_type = 1,
  line_width = 1, add_points = T, ...)
}
\arguments{
\item{data}{dataset.}

\item{by_var}{by variable.}

\item{group_var}{grouping variable.}

\item{time_var}{time variable.}

\item{conc_var}{response variable.}
}
\value{
hormLong object.
}
\description{
Create longitudinal plot with smoothers (under development)
}
\examples{
result <- hormBaseline(data=hormLynx, criteria=2, by_var='AnimalID, Hormone', time_var='Date', conc_var='Conc' )
}

