% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hormSumTable}
\alias{hormSumTable}
\title{Summarise results from baseline calculation.}
\usage{
hormSumTable(x, num_decimals = 2)
}
\arguments{
\item{x}{hormLong object (produced from hormBaseline) [required]}

\item{num_decimals}{number of decimal places [default = 2]}
}
\value{
nothing  Produces a file saved to current working directory
}
\description{
Summarise results from baseline calculation.
}
\examples{
result <- hormBaseline(data=hormone, by_var='sp, sex, id', time_var='date', conc_var='conc' )
hormSumTable(result)
}

