% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hormSumTable.R
\name{hormSumTable}
\alias{hormSumTable}
\title{Summarise results from baseline calculation.}
\usage{
hormSumTable(x, num_decimals = 2)
}
\arguments{
\item{x}{hormLong object (produced from hormBaseline) [required]}

\item{num_decimals}{number of decimal places [default = 2]}
}
\value{
nothing  Produces a file saved to current working directory
}
\description{
Summarise results from baseline calculation.
}
\examples{
result <- hormBaseline(data=hormLynx, criteria=2, by_var='AnimalID, Hormone', time_var='datetime', conc_var='Conc' )
hormSumTable(result)
}

