% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hormSumTable}
\alias{hormSumTable}
\title{Summarise results from baseline calculation.}
\usage{
hormSumTable(x, file_type = "rtf", num_decimals = 2)
}
\arguments{
\item{x}{hormLong object (produced from hormBaseline)}

\item{file_type}{choose either rtf or csv}

\item{num_decimals}{number of decimal places}
}
\value{
hormLong object.
}
\description{
Summarise results from baseline calculation.
}
\examples{
result <- hormBaseline(data=hormone, by_var='sp, sex, id', time_var='date', conc_var='conc' )
hormSumTable(result)
}

