% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hormWrite}
\alias{hormWrite}
\title{Write data in hormLong object to file}
\usage{
hormWrite(x, filename, file_type = "csv", ...)
}
\arguments{
\item{x}{hormLong object (produced from hormBaseline)}

\item{filename}{name of file to be saved. Unless path is specified, file will be
saved in your current working directory}

\item{file_type}{determines type of file (e.g. csv, tab-delimited)}

\item{...}{options for write.csv}
}
\value{
nothing
}
\description{
Write data in hormLong object to file
}
\examples{
result <- hormBaseline(data=hormone, by_var='sp, sex, id', time_var='date', conc_var='conc' )
hormPlot(result)
hormWrite(result, file_out='data_out.csv', file_type='csv' )
}

