% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hormHelpers.R
\name{plotLines}
\alias{plotLines}
\title{Helper function for plotting lines using break_cutoff so lines are not joined if large temporal gap}
\usage{
plotLines(d = ds_sub, c = conc_var, t = time_var, color = "black")
}
\arguments{
\item{d}{dataset to plot (usually ds_sub)}

\item{c}{conc_var (y-axis)}

\item{t}{time_var (x-axis)}

\item{color}{line colour [default='black']}
}
\value{
nothing
}
\description{
Helper function for plotting lines using break_cutoff so lines are not joined if large temporal gap
}

